requirements.txt
git+https://github.com/openai/CLIP.git
discord.py==1.7.3
gptj==3.0.7
Pillow==9.0.0
python-dotenv==0.19.2
torch==1.10.1
numpy~=1.22.0